JSON that contains Jersey CI Bank Holidays - useful for identifying is a date is a bank holiday
This project exists as I need to perform validation on dates for a
client and make sure a bank holiday was not selected.    I found some
UK bank holidays https://www.gov.uk/bank-holidays.json that could be
used as part of some javascript validation on when a date enters.  This
project transforms that data from gov.uk to have liberation day
included.